printing inputs and labels

+++ code/gan.py
@@ -102,12 +102,12 @@ def img_reader():
 
 
 
-        depth_arr = np.array(d_image)#.flatten()
-        rgb_arr = np.array(rgb_image)#.flatten()
+        depth_arr = np.array(d_image, dtype=float)#.flatten()
+        rgb_arr = np.array(rgb_image, dtype=float)#.flatten()
 
         # Normalize between between -1 and 1
-        rgb_norm = rgb_arr / float(np.max(np.abs(rgb_arr), axis=0)) # (2 * (rgb_arr - np.max(rgb_arr))) / (-np.ptp(rgb_arr) - 1)
-        depth_norm = depth_arr / float(np.max(np.abs(depth_arr), axis=0)) # (2 * (depth_arr - np.max(depth_arr))) / (-np.ptp(depth_arr) - 1)
+        rgb_norm = rgb_arr / np.max(np.abs(rgb_arr), axis=0) # (2 * (rgb_arr - np.max(rgb_arr))) / (-np.ptp(rgb_arr) - 1)
+        depth_norm = depth_arr / np.max(np.abs(depth_arr), axis=0) # (2 * (depth_arr - np.max(depth_arr))) / (-np.ptp(depth_arr) - 1)
 
         d_image = Image.fromarray((depth_norm * 255).astype('uint8'))
         d_image.convert('L').save('/mnt/results/depth.jpg')
